separate Movie and Director into separate directories
