+++ src/main/java/com/artem/movieViewer/director/repository/api/DirectorRepository.java
@@ -1,6 +1,6 @@
-package com.artem.movieViewer.movie.repository.api;
+package com.artem.movieViewer.director.repository.api;
 
-import com.artem.movieViewer.movie.entity.Director;
+import com.artem.movieViewer.director.entity.Director;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
